// src/enums.ts
export enum RuneType {
	Fire = 'Fire',
	Earth = 'Earth',
	Water = 'Water',
	Dark = 'Dark',
	Wild = 'Wild',
	Wind = 'Wind',
}

export enum CharacterName {
	Volgr = 'Volgr',
	Maelling = 'Maelling',
	Lydana = 'Lydana',
	Kattly = 'Kattly',
	Briya = 'Briya',
	Felerion = 'Felerion',
	Error = 'Error',
}

export enum ActionType {
	AddRune = 'add-rune',
	RemoveRune = 'remove-rune',
	ResetCharacter = 'reset-character',
}